Functions/functions/Image_Operations/requirements.txt
appwrite
asyncio
aiohttp
Pillow
numpy
tensorflow
opencv-python-headless